Revert "DAOS-12949 pipeline: add layout version to OID (#11903)"
This reverts commit be48a6714fce24ddfef9580632cfde5721e1b478.

+++ src/object/obj_layout.c
@@ -19,7 +19,7 @@ obj_pl_grp_idx(uint32_t layout_gl_ver, uint64_t hash, uint32_t grp_nr)
 
 /* Choose EC start offset within the group. */
 int
-obj_ec_grp_start(uint16_t layout_gl_ver, uint64_t hash, uint32_t grp_size)
+obj_ec_grp_start(uint32_t layout_gl_ver, uint64_t hash, uint32_t grp_size)
 {
 	if (layout_gl_ver == 0)
 		return 0;
@@ -30,7 +30,7 @@ obj_ec_grp_start(uint16_t layout_gl_ver, uint64_t hash, uint32_t grp_size)
 
 /* Generate the object layout */
 int
-obj_pl_place(struct pl_map *map, uint16_t layout_gl_ver, struct daos_obj_md *md,
+obj_pl_place(struct pl_map *map, uint32_t layout_gl_ver, struct daos_obj_md *md,
 	     unsigned int mode, uint32_t rebuild_ver, struct daos_obj_shard_md *shard_md,
 	     struct pl_obj_layout **layout_pp)
 {
